feature/error handling on announcements model

+++ app/models/Announcements.js
@@ -1,4 +1,5 @@
 const util = require('../data/util');
+const output = util.output;
 const currentNumberOfAnnouncements = util.currentNumberOfAnnouncements;
 const existedAnnouncements = require('../data/existedAnnouncements');
 const assignAnnouncementNumber = util.assignAnnouncementNumber;
@@ -7,14 +8,23 @@ assignAnnouncementNumber(existedAnnouncements, currentNumberOfAnnouncements);
 
 module.exports = async (client) => {
     try {
-        const announcements = client.db("ListamDB").collection("announcements");
-        await announcements.drop();
-        // console.log('Announcement table dropped succesfully!')
+        const db = client.db("ListamDB");
+        db.listCollections({name: 'announcements'})
+            .next(function(err, exist) {
+                if (exist) {
+                    const announcements = db.collection("announcements")
+                    announcements.drop()
+                    .then(() => console.log('Announcement table dropped succesfully!'))
+                    .catch(error => output(error));
+                }
+        });
 
         const options = { ordered: true };
-        const result = await announcements.insertMany(existedAnnouncements, options);
-
-        // console.log(`${result.insertedCount} documents inserted!`);
+        return announcements.insertMany(existedAnnouncements, options)
+        .then(result => console.log(`${result.insertedCount} documents inserted!`))
+        .catch(error => output(error));
+    } catch(error) {
+        output(error);
     } finally {
         // client.close();
     }
